Factor out `module-info` module
